Don't expose Postgres port in Docker
This avoids `docker-compose up` failing if ran on a machine that has Postgres
running locally on the same port. As the other containers access Postgres via
Docker, they'll continue working regardless if we're exposing its port or not.
Also, if we need to access the DB for whatever reason, we can get the random IP
assigned by Docker using `docker inspect` and use it instead.