Add make setup to the docs
This was not essential until we need to use the seed command, otherwise
no data would be saved in the database since `territory_id` must be a
valid foreign key. This simple commit documents in the CONTRIBUTING.md
the need to run `make setup` (since `make setup` calls `make seed` too).