Remove Scrapy Contracts from existing spiders
The initial idea of using Contracts was to have a quick
(for experienced developers) and beginner-friendly way of writing
Scrapy tests, exempting developers from having to write test cases,
creating fixtures and maintaining them up-to-date.

The downside is that these contracts commonly raise false positives,
when the scraper is working correctly but the contract fails,
temporarily or not, for other reasons.

Since the number of false positives grow in the past months without
maintenance, I am currently removing them. We should think about
alternatives in the future.

Following discussion from
https://github.com/okfn-brasil/diario-oficial/issues/44.