scrapy-deltafetch: avoid crawl the same URL
Adds the scrapy-deltafetch middleware. It allow us to drop item which
already has been crawled. It uses the URL to identify if an item has
already been download. Even if it can causes issues if the same URL
points to a different data in the future, it sounds reasonable to add it
to avoid waste of resources. Furthermore, it is always possible to set
the middleware to reset and allow all the item to be crawled again.

https://github.com/okfn-brasil/diario-oficial/issues/86

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>