Add Recife's Gazette Attachment flow
This spider has three constraints that had to be dealt with:

1. The gazette's document is generated through a request that makes it
   available in a server cache for a period of time and only for the
   session which requested it. That's why cookies were enabled;
2. The spider needs cookiejars to be able to crawl multiple dates
   concurrently. Because of the first constraint the cookies must be
   passed through until the gazette's document request. Hence the new
   GazetteFilesPipeline;
3. It blocks some requests if multiple requests are made in a short
   time frame. AUTOTHROTTLE_ENABLED seems to have solved it, though.
   It also has the added benefit of not putting unnecessary burden on
   the server.

... and one that hasn't been solved ...

Because of the second constraint and the new pipeline, the requests
that download the documents are not passed through middlewares. Thus,
DeltaFetch cannot check for the request's previous occurrence. In other
words, the spider can download the same gazette document more than once
if it is executed more than once.