Adds QueridoDiarioFilesPipeline
Adds the QueridoDiarioFilesPipeline class. A specialization of the scrapy
FilesPipeline. The new class will allow the users to store in the
backends available in the scrapy class but changing the directory where
the file will be stored. The current version will keep all the gazettes
files from the same date in the same directory.

Related to issue https://github.com/okfn-brasil/diario-oficial/issues/165

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>