Added two new variable. POSTGRES_HOST_PORT and POSTGRES_DOCKER_PORT to control easily the ports used in postgres. If you have postgres installed in you local machine, it necessary just to change the value of POSTGRES_HOST_PORT to avoid port conflict.
i've just make some improvements to make the docker-compose.yml more readable. The container_name sets the actual name of the container when it runs, rather than letting Docker Compose generate it. Also create a network called production-network to provide complete isolation for containers. The service processing now is build automatic if docker not found the image. In addition, i added the tag command. Just change the this line to run the scrapy that you want

make some improvements to make more easily to undestand the Dockerfile. Also added one entrypoint to image.

change the bash command because i change the WORKDIR in Dockerfile of processing service

added to gitignore the vim temporary files

added the image tag and remove one unnecessary line

create a dockerignore to send only the necessary files to image

changed the path of code to keep the path of others files

clean up the apt cache by removing /var/lib/apt/lists to reduce the image size

no-install-recommends reduce the size of image by avoiding installing packages that aren’t technically dependencies but are recommended to be installed alongside packages

remove the commented lines

remove the new WORKDIR and CMD

return the makefile to use the previous path in container