Fix rj_campos_goytacazes spider
The spider would not run because `gazette_text` is `None` for the first
element yielded by the for loop, so calling `startswith` on it would
throw an exception.

This commit fixes the problem by setting `gazette_text` to an empty
string in case it's `None` and using the `date_re` value to skip the
element in case its text does not match what is expected.