Remove scrapy-deltafetch middleware
This middleware was introduced to avoid spiders to request data that was
already requested before, reducing the amount of requests made. However,
it is not certain that it will reach this goal considering that some
requests may be the same but the content returned different. A better
solution to reduce the number of requests may be necessary, but we will
only know for sure when we start running all the spiders daily.