Inserted Guidelines to maintainers
I could open an issue to discuss this, but I have adopted a philosophy of starting a discussion with a proposal.

I created this chapter to propose an organization for the project maintainers. The idea is that it is not legislation with multiple responsibilities, but a guide that facilitates communication and reduces rework.

I do this when we invite @giuliocc to compose our team of maintainers. :1st_place_medal:  Thus, we are growing and creating a working model to prevent distress to people as generous as Querido Diario maintainers.

"There is a criteria I call change latency, which is the round-trip time from identifying a problem to testing a solution. The faster the better. If maintainers cannot respond to pull requests as rapidly as people expect, they're not doing their job (or they need more hands)." (Peter Hintjens)
https://hintjens.gitbooks.io/social-architecture/content/chapter4.html

I would like opinions: @jvanz @rennerocha