Processed column
Adds a new column in the Gazette model. This column should be used by
the processing pipeline to find which gazettes is pending to be ingested
by the system. This is not the best solution to this, a queue system
would be a better approach for instance. But for the current state of the
project this will be enough.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>