File processed flag (#306)
* .gitignore: sqlite database file

Adds the file used by SQLite to store the gazette data in the .gitignore

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>

* SQLite client

Updates the container used to run the spider installing the package
necessary to access the SQLite database. Furthermore, the "sql" make
target now open the sqlite command line. Allowing the user to interact
with the database if necessary.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>

* Processed column

Adds a new column in the Gazette model. This column should be used by
the processing pipeline to find which gazettes is pending to be ingested
by the system. This is not the best solution to this, a queue system
would be a better approach for instance. But for the current state of the
project this will be enough.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>