Store all files information available for a date
Some cities have more than one file for the same gazette edition. We
need to store uploaded file information (like URL, path and checksum)
for all of them. We were storing just the information of the first file
available (which is ok for the majority of the cities that have only one
file per date), but fails if we have more files available.