Misleading error message when write in the database fails.
The integrity error raised by SQLAlchemy is launched not only when the
item is already in the database. For instance, it can also be launched
when the territories table is missing some territory ID. Thus, this
commit updates the error message and add the exception in the log
message. Thus, the users will be able to detect which kind of error
they are facing.

Signed-off-by: José Guilherme Vanz <jvanz@jvanz.com>