Do not raise SQLAlchemyError as errors in job logs
Some pages has different URLs for the same file (Vinhedo-SP for
example). Because of this, the `status` of the file is not `uptodated`
and we try to insert it again in the database (with the same checksum).
These exceptions are logged as errors, so the monitors fail. However, we
can ignore these errors, just adding a warning log.